refactor : 컴포넌트 위치 구조화 진행
