+++ components/features/blog/detail/PostDetailContent.tsx
