+++ components/layouts/Footer.jsx
@@ -3,7 +3,7 @@ export default function Footer() {
     <footer className="border-t">
       <div className="container flex h-14 items-center justify-center">
         <p className="text-muted-foreground text-sm">
-          Built with Next.js, Tailwind CSS and shadcn/ui
+          백과의 개인 블로그 입니다. Contact : ksu9801@gmail.com
         </p>
       </div>
     </footer>
